Move utility functions into util folder
