+++ src/util/functions.ts
